progress bar fixed

+++ Assets/Scripts/Features/DominationBar/DominationManagerDataHandler.cs
@@ -2,6 +2,6 @@ using System;
 
 public static class DominationManagerDataHandler 
 {
-    public static event Action<int> OnUpdateDominationBar;
-    public static void UpdateDominationBar(int candleChangeNb) => OnUpdateDominationBar?.Invoke(candleChangeNb);
+    public static event Action<float> OnUpdateDominationBar;
+    public static void UpdateDominationBar(float candleChangeNb) => OnUpdateDominationBar?.Invoke(candleChangeNb);
 }
